four commit

+++ src/main/java/org/exampleProjectOne/util/DBHelper.java
@@ -25,21 +25,22 @@ public class DBHelper {
         return dbHelper;
     }
 
+    private static SessionFactory getSessionFactory () {
+        Configuration configuration = DBHelper.getInstance().getConfiguration();
+        StandardServiceRegistryBuilder builder = new StandardServiceRegistryBuilder();
+        builder.applySettings(configuration.getProperties());
+        ServiceRegistry serviceRegistry = builder.build();
+        return configuration.buildSessionFactory(serviceRegistry);
+    }
+
     public Connection getConnection() {
         try {
-            DriverManager.registerDriver((Driver) Class.forName("com.mysql.jdbc.Driver").newInstance());
-
-            StringBuilder url = new StringBuilder();
-
-            url.
-                    append("jdbc:mysql://").        //db type
-                    append("localhost:").           //host name
-                    append("3306/").                //port
-                    append("mysql?").          //db name
-                    append("user=root&").          //login
-                    append("password=admin65");       //password
-
-            Connection connection = DriverManager.getConnection(url.toString());
+            DriverManager.registerDriver((Driver) Class.forName(
+                    PropertyRead.readProperty("db.driver")).newInstance());
+            Connection connection = DriverManager.getConnection(
+                    PropertyRead.readProperty("db.url"),
+                    PropertyRead.readProperty("db.userName"),
+                    PropertyRead.readProperty("db.passwordDB"));
             return connection;
         } catch (SQLException | InstantiationException | IllegalAccessException | ClassNotFoundException e) {
             e.printStackTrace();
@@ -50,53 +51,56 @@ public class DBHelper {
     public Configuration getConfiguration() {
         Configuration configuration = new Configuration();
         configuration.addAnnotatedClass(User.class);
-        configuration.setProperty("hibernate.dialect", "org.hibernate.dialect.MySQL57Dialect");
-        configuration.setProperty("hibernate.connection.driver_class", "com.mysql.jdbc.Driver");
-        configuration.setProperty("hibernate.connection.url", "jdbc:mysql://localhost:3306/mysql");
-        configuration.setProperty("hibernate.connection.username", "root");
-        configuration.setProperty("hibernate.connection.password", "admin65");
-        configuration.setProperty("hibernate.show_sql", "true");
-        configuration.setProperty("hibernate.hbm2ddl.auto", "create");
+        configuration.setProperty("hibernate.dialect", PropertyRead.readProperty("db.dialect"));
+        configuration.setProperty("hibernate.connection.driver_class", PropertyRead.readProperty("db.driver"));
+        configuration.setProperty("hibernate.connection.url", PropertyRead.readProperty("db.url"));
+        configuration.setProperty("hibernate.connection.username", PropertyRead.readProperty("db.userName"));
+        configuration.setProperty("hibernate.connection.password", PropertyRead.readProperty("db.passwordDB"));
+        configuration.setProperty("hibernate.show_sql", PropertyRead.readProperty("db.show"));
+        configuration.setProperty("hibernate.hbm2ddl.auto", PropertyRead.readProperty("db.typeCreate"));
         return configuration;
     }
 
-    public static void initHibernate(SessionFactory sessionFactory) {
-        User userOne = new User("Admin", "admin", 1L);
-        User userTwo = new User("AdminTwo", "adminTwo", 2L);
-        try (UserHibernateDAO userHibernateDAO = new UserHibernateDAO(sessionFactory.openSession());) {
-            userHibernateDAO.addUser(userOne);
-            userHibernateDAO.addUser(userTwo);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    public static void initializeMysql() throws Exception {
-        DriverManager.registerDriver((Driver) Class.forName("com.mysql.cj.jdbc.Driver").newInstance());
-        StringBuilder url = new StringBuilder();
-
-        url.
-                append("jdbc:mysql://").        //db type
-                append("localhost:").           //host name
-                append("3306/").                //port
-                append("mysql?").          //db name
-                append("user=root&").          //login
-                append("password=admin65");       //password
-
-        Connection connection = DriverManager.getConnection(url.toString());
-        Statement stmt = connection.createStatement();
-        stmt.executeUpdate("DROP TABLE IF EXISTS users");
-        stmt.execute("create table if not exists users (id bigint auto_increment, name varchar(256), password varchar(256), age bigint, primary key (id))");
-        stmt.close();
-        connection.close();
-
+    public static void initDataResources(){
         User userOne = new User("Admin", "admin", 1L);
         User userTwo = new User("AdminTwo", "adminTwo", 2L);
 
-        try (Connection connectionTry = DriverManager.getConnection(url.toString())) {
-            UserJdbcDAO userJdbcDAO = new UserJdbcDAO(connectionTry);
-            userJdbcDAO.addUser(userOne);
-            userJdbcDAO.addUser(userTwo);
+        if (PropertyRead.readProperty("db.host").equals("hibernate")) {
+            UserHibernateDAO.setSessionFactory(getSessionFactory());
+            try (UserHibernateDAO userHibernateDAO = new UserHibernateDAO(
+                    UserHibernateDAO.getSessionFactory().openSession())) {
+                userHibernateDAO.addUser(userOne);
+                userHibernateDAO.addUser(userTwo);
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        } else if (PropertyRead.readProperty("db.host").equals("jdbc")) {
+            try {
+                DriverManager.registerDriver((Driver) Class.forName(
+                        PropertyRead.readProperty("db.driver")).newInstance());
+
+                try (Connection connectionTry = DriverManager.getConnection(
+                        PropertyRead.readProperty("db.url"),
+                        PropertyRead.readProperty("db.userName"),
+                        PropertyRead.readProperty("db.passwordDB"))) {
+                    Statement stmt = connectionTry.createStatement();
+                    stmt.executeUpdate("DROP TABLE IF EXISTS users");
+                    stmt.execute("create table if not exists users (id bigint auto_increment, name varchar(256), password varchar(256), age bigint, primary key (id))");
+                    stmt.close();
+                }
+
+                try (Connection connectionTry = DriverManager.getConnection(
+                        PropertyRead.readProperty("db.url"),
+                        PropertyRead.readProperty("db.userName"),
+                        PropertyRead.readProperty("db.passwordDB"))) {
+                    UserJdbcDAO userJdbcDAO = new UserJdbcDAO(connectionTry);
+                    userJdbcDAO.addUser(userOne);
+                    userJdbcDAO.addUser(userTwo);
+                }
+            } catch (SQLException | InstantiationException | IllegalAccessException | ClassNotFoundException e) {
+                e.printStackTrace();
+                throw new IllegalStateException();
+            }
         }
     }
